04-drug-interaction-ml/pyproject.toml
[project]
name = "drug-interaction-ml"
version = "0.1.0"
description = "Drug-Drug Interaction Prediction Pipeline with ML deployment, drift detection, and monitoring"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "DrugInteractionML Team" }]

dependencies = [
    "snowflake-connector-python>=3.6.0",
    "boto3>=1.34.0",
    "scikit-learn>=1.4.0",
    "xgboost>=2.0.0",
    "rdkit-stubs>=0.7",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "apache-airflow>=2.8.0",
    "pydantic>=2.6.0",
    "mlflow>=2.10.0",
    "shap>=0.44.0",
    "scipy>=1.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "moto[sagemaker,sns,cloudwatch,stepfunctions]>=5.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "ANN"]
ignore = ["ANN101", "ANN102", "ANN401"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
